// pages/api/types/incoming/garmin-activity.ts
interface GarminActivity {
    activityId: number;
    activityName: string;
    description?: any;
    startTimeLocal: string;
    startTimeGMT: string;
    activityType: ActivityType;
    eventType: EventType;
    comments?: any;
    parentId?: any;
    distance: number;
    duration: number;
    elapsedDuration: number;
    movingDuration: number;
    elevationGain: number;
    elevationLoss: number;
    averageSpeed: number;
    maxSpeed: number;
    startLatitude: number;
    startLongitude: number;
    hasPolyline: boolean;
    ownerId: number;
    ownerDisplayName: string;
    ownerFullName: string;
    ownerProfileImageUrlSmall: string;
    ownerProfileImageUrlMedium: string;
    ownerProfileImageUrlLarge: string;
    calories: number;
    averageHR: number;
    maxHR: number;
    averageRunningCadenceInStepsPerMinute?: any;
    maxRunningCadenceInStepsPerMinute?: any;
    averageBikingCadenceInRevPerMinute?: any;
    maxBikingCadenceInRevPerMinute?: any;
    averageSwimCadenceInStrokesPerMinute?: any;
    maxSwimCadenceInStrokesPerMinute?: any;
    averageSwolf?: any;
    activeLengths?: any;
    steps?: any;
    conversationUuid?: any;
    conversationPk?: any;
    numberOfActivityLikes?: any;
    numberOfActivityComments?: any;
    likedByUser?: any;
    commentedByUser?: any;
    activityLikeDisplayNames?: any;
    activityLikeFullNames?: any;
    activityLikeProfileImageUrls?: any;
    requestorRelationship?: any;
    userRoles: string[];
    privacy: Privacy;
    userPro: boolean;
    courseId?: any;
    poolLength?: any;
    unitOfPoolLength?: any;
    hasVideo: boolean;
    videoUrl?: any;
    timeZoneId: number;
    beginTimestamp: number;
    sportTypeId: number;
    avgPower?: any;
    maxPower?: any;
    aerobicTrainingEffect: number;
    anaerobicTrainingEffect: number;
    strokes?: any;
    normPower?: any;
    leftBalance?: any;
    rightBalance?: any;
    avgLeftBalance?: any;
    max20MinPower?: any;
    avgVerticalOscillation?: any;
    avgGroundContactTime?: any;
    avgStrideLength?: any;
    avgFractionalCadence?: any;
    maxFractionalCadence?: any;
    trainingStressScore?: any;
    intensityFactor?: any;
    vO2MaxValue?: any;
    avgVerticalRatio?: any;
    avgGroundContactBalance?: any;
    lactateThresholdBpm?: any;
    lactateThresholdSpeed?: any;
    maxFtp?: any;
    avgStrokeDistance?: any;
    avgStrokeCadence?: any;
    maxStrokeCadence?: any;
    workoutId?: any;
    avgStrokes?: any;
    minStrokes?: any;
    deviceId: number;
    minTemperature?: any;
    maxTemperature?: any;
    minElevation: number;
    maxElevation: number;
    avgDoubleCadence?: any;
    maxDoubleCadence?: any;
    summarizedExerciseSets?: any;
    maxDepth?: any;
    avgDepth?: any;
    surfaceInterval?: any;
    startN2?: any;
    endN2?: any;
    startCns?: any;
    endCns?: any;
    summarizedDiveInfo: SummarizedDiveInfo;
    activityLikeAuthors?: any;
    avgVerticalSpeed?: any;
    maxVerticalSpeed: number;
    floorsClimbed?: any;
    floorsDescended?: any;
    manufacturer: string;
    diveNumber?: any;
    locationName: string;
    bottomTime?: any;
    lapCount: number;
    endLatitude: number;
    endLongitude: number;
    minAirSpeed?: any;
    maxAirSpeed?: any;
    avgAirSpeed?: any;
    avgWindYawAngle?: any;
    minCda?: any;
    maxCda?: any;
    avgCda?: any;
    avgWattsPerCda?: any;
    flow?: any;
    grit?: any;
    jumpCount?: any;
    caloriesEstimated?: any;
    caloriesConsumed?: any;
    waterEstimated: number;
    waterConsumed?: any;
    maxAvgPower_1?: any;
    maxAvgPower_2?: any;
    maxAvgPower_5?: any;
    maxAvgPower_10?: any;
    maxAvgPower_20?: any;
    maxAvgPower_30?: any;
    maxAvgPower_60?: any;
    maxAvgPower_120?: any;
    maxAvgPower_300?: any;
    maxAvgPower_600?: any;
    maxAvgPower_1200?: any;
    maxAvgPower_1800?: any;
    maxAvgPower_3600?: any;
    maxAvgPower_7200?: any;
    maxAvgPower_18000?: any;
    excludeFromPowerCurveReports?: any;
    totalSets?: any;
    activeSets?: any;
    totalReps?: any;
    minRespirationRate?: any;
    maxRespirationRate?: any;
    avgRespirationRate?: any;
    trainingEffectLabel?: any;
    activityTrainingLoad?: any;
    avgFlow?: any;
    avgGrit?: any;
    minActivityLapDuration: number;
    avgStress?: any;
    startStress?: any;
    endStress?: any;
    differenceStress?: any;
    maxStress?: any;
    aerobicTrainingEffectMessage: string;
    anaerobicTrainingEffectMessage: string;
    splitSummaries: any[];
    hasSplits: boolean;
    moderateIntensityMinutes: number;
    vigorousIntensityMinutes: number;
    hasSeedFirstbeatProfile?: any;
    favorite: boolean;
    decoDive?: any;
    purposeful: boolean;
    pr: boolean;
    autoCalcCalories: boolean;
    parent: boolean;
    atpActivity: boolean;
    manualActivity: boolean;
    elevationCorrected: boolean;
}

interface SummarizedDiveInfo {
    weight?: any;
    weightUnit?: any;
    visibility?: any;
    visibilityUnit?: any;
    surfaceCondition?: any;
    current?: any;
    waterType?: any;
    waterDensity?: any;
    summarizedDiveGases: any[];
    totalSurfaceTime?: any;
}

interface Privacy {
    typeId: number;
    typeKey: string;
}

interface EventType {
    typeId: number;
    typeKey: string;
    sortOrder: number;
}

interface ActivityType {
    typeId: number;
    typeKey: string;
    parentTypeId: number;
    sortOrder: number;
    isHidden: boolean;
    restricted: boolean;
}